refactor tests
[ghstack-poisoned]
