+++ test/jit/test_profiler.py
@@ -1,0 +1,25 @@
+import os
+import sys
+
+import torch
+
+# Make the helper files in test/ importable
+pytorch_test_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
+sys.path.append(pytorch_test_dir)
+from torch.testing._internal.jit_utils import JitTestCase
+
+if __name__ == '__main__':
+    raise RuntimeError("This test file is not meant to be run directly, use:\n\n"
+                       "\tpython test/test_jit.py TESTNAME\n\n"
+                       "instead.")
+
+class TestProfiler(JitTestCase):
+    def setUp(self):
+        self.prev_exec = torch._C._jit_set_profiling_executor(True)
+        self.prev_profiling = torch._C._jit_set_profiling_mode(True)
+        self.inline_autodiff = torch._C._debug_set_autodiff_subgraph_inlining(False)
+
+    def tearDown(self):
+        torch._C._jit_set_profiling_executor(self.prev_exec)
+        torch._C._jit_set_profiling_mode(self.prev_profiling)
+        torch._C._debug_set_autodiff_subgraph_inlining(self.inline_autodiff)
