+++ test/test_jit_fuser_te.py
@@ -21,7 +21,7 @@ torch._C._jit_set_profiling_mode(True)
 from torch.testing._internal.common_utils import run_tests, IS_SANDCASTLE, ProfilingMode, GRAPH_EXECUTOR, \
     enable_profiling_mode_for_profiling_tests
 from torch.testing._internal.jit_utils import JitTestCase, _inline_everything, \
-    RUN_CUDA, RUN_CUDA_HALF, RUN_CUDA_MULTI_GPU
+    RUN_CUDA, RUN_CUDA_HALF, RUN_CUDA_MULTI_GPU, warmup_backward
 
 from textwrap import dedent
 from itertools import product, permutations
@@ -37,27 +37,12 @@ def strip_profiling_nodes(nodes):
     profiling_opcodes = set(['prim::BailoutTemplate', 'prim::BailOut'])
     return [n for n in nodes if n.kind() not in profiling_opcodes]
 
-
-def warmup_backward(f, *args):
-    profiling_count = 2
-    results = []
-    for i in range(profiling_count):
-        if len(args) > 0:
-            r = torch.autograd.grad(f, *args)
-            results.append(r)
-        else:
-            f.backward(retain_graph=True)
-
-    return results
-
-
 def warmup_forward(f, *args):
     profiling_count = 2
     for i in range(profiling_count):
         results = f(*args)
 
     return results
-
 
 class TestTEFuser(JitTestCase):
     def setUp(self):
